requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.1
Django==5.1.6
django-debug-toolbar==5.0.1
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.2
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.4
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.18.3
